added and fixed bugs with deregister, added buttons for clean blood

+++ api/handleDeregisterDonor.py
@@ -1,7 +1,7 @@
-#This Function should remove the username to the database of people from which we can potentially collect blood
+# This Function should remove the username to the database of people from which we can potentially collect blood
 def handleDeregisterDonor(name, db):
-   for item in db:
-      if (item['name'] == name):
-         db.remove(item)
-   return "{ 'value': 'success' }", db
-
+    for item in db:
+        if (item['type'] == 'donor'):
+            if (item['name'] == name):
+                db.remove(item)
+    return '{ "value": "success" }', db
